requirements.txt
fastapi==0.88.0
uvicorn==0.18.3
psycopg2-binary==2.9.3
sqlalchemy==1.4.41
python-dotenv==0.20.0
passlib==1.7.4
python-jose==3.3.0
python-multipart==0.0.5
jinja2==3.1.2